Update Add properties to car class
Add properties to car class
You will add more features to the Car class and a few properties that mimic a real car. A real car has endless features, but for now, you will only add the car's color, model, and year of manufacture.

Instructions

Add a string property color to the Car class for the color of a car.
Add a string property model to the Car class for the model of a car.
Add an integer property year to the Car class to the year a car was made.