Create Creating a constructor for car class
Creating a constructor for car class
It is important to be able to set values for a class’s properties, in order to use them effectively. You will create a Constructor for the Car class and ensure that its properties are set to some values.

Remember: The constructor of a class will always be called when we create an instance.

Instructions

Create a constructor method for the Car class that takes no parameters.
Inside the constructor, set the color property to be a String "red"
Inside the constructor, set the year to be an Integer 2019
Print out the color property of the myCar object instance.