Create Adding parameters to car class constructor
Adding parameters to car class constructor
To use classes dynamically, the class constructor needs to take parameters. You will update the Car class and make the constructor receive parameters, ensuring each instance you create can have different property values.

Instructions

Update the constructor for the Car class to take one more parameter: an int year.
Set the year inside the constructor to the matching constructor parameters.
Pass the values "blue" as a String, "corolla" as a String, and 2022 as an int, to the Car constructor when creating myCar.