Create Adding class methods with return types
Adding class methods with return types
Cars have a performance metric called miles per gallon that tells the driver the fuel consumption rate. You will create a method to calculate the miles per gallon called calculateMPG. The calculateMPG method uses the number of miles driven and the amount of gallons used to calculate and return the correct miles per gallon unit.

Instructions

Create a calculateMPG method that returns an int type.
The calculateMPG method should take two parameters an int milesDriven and an int gallonsUsed.
The calculateMPG should return the division value of milesDriven divided by gallonsUsed.
Use myCar object instance to call and print the value of the calculateMPG method, with the parameter values 180 and 20.