Create Implementing abstract methods
Implementing abstract methods
You will change the Toyota and Mercedes classes to give each car its implementation for the horn sounding. For uniqueness, you will create a version of the method for sounding the horn in both the Toyota and Mercedes classes.

NOTE: An abstract method must be implemented when inherited from an Abstract class, within the subclass.

Instructions

Implement the soundHorn method in the Toyota class to print "sounds like a toyota".
Implement the soundHorn method inside the Mercedes class to print "sounds like a mercedes".
Inside the main method, call the soundHorn method on the myToyota instance and the soundHorn method on the myMercedes instance.